C# sample app - fix slow textbox update
AppendText is vastly more efficient than the current implementation.

I'm assuming that the current code to limit the contents of the textbox
to 100 lines (not 6 as stated in the comment) was done on the assumption
that this would make updating the textbox faster, but it actually makes
things much worse. I can't see any other point in dong this.