Post to SynchronizationContext instead of using tasks
Posting to the synchronization context instead of running a task is much
more efficient but achieves the same effect.

Also I don't know that there's any guarantee that the tasks will execute
in the order they're queued to the task scheduler, so it's possible that
the application will received events out of sequence. Whereas posting to
the sync context does ensure in-sequence execution.