Use private object for locking
EClient uses lock (this) to prevent concurrent write access to the
socket. But EClient is public, and using a public object for locking is
bad practice (because any code could acquire the lock).

This commit uses a private object for locking. It also moves the locking
to where it's actually needed, the current locking point being too
coarse.