Allow wrapper handlers to per-instance methods.
The current way of invoking the wrapper methods, when handling messages, allows the wrapper class to be subclassed by the user. However, it does not allow the user to programmably substitute a per-instance method of an already created wrapper object. This pull request is aimed at fixing that. It should be backward compatible.